It is uploaded via Jupyter Notebook.
This Sample code of Support Vector Machine (SVM)

SVM ----> Support Vector Machines

SVM is a supervised learning algorithm which is mainly designed to perform segregation of data by drawing a decision boundary.

SVM can be used for both classification and regression problem statements.

SVM is a very good classification and not recommended for regression.

Decision boundary is a straight line which segregate/grouping the data into multiple clusters, it is placed at maximum distance from closest point of each cluster.

The data points each clusters which helps us to identify the decision boundary is called as Support vectors. So , these closest data points are very nearer to decision boundary.

Decision boundary we also called as Hyper plane.

The two parallel lines on either side of decision boundary by passing through the support vectors is called as Margins.

The distance between the margins which are placed on either side of decision boundary are called as marginal distance.